+++ src/bda/plone/shop/browser/actions.py
@@ -4,9 +4,9 @@ from zope.interface import noLongerProvides
 from Products.Five.browser import BrowserView
 from bda.plone.orders.interfaces import IVendor
 from bda.plone.orders import permissions
-from ..interfaces import IBuyable
-from ..interfaces import IPotentiallyBuyable
-from .. import message_factory as _
+from bda.plone.shop.interfaces import IBuyable
+from bda.plone.shop.interfaces import IPotentiallyBuyable
+from bda.plone.shop import message_factory as _
 
 
 class EnableDisableFeature(BrowserView):