+++ src/bda/plone/shop/browser/controlpanel.py
@@ -3,9 +3,9 @@ from zope.component import getUtility
 from zope.dottedname.resolve import resolve
 from plone.app.registry.browser import controlpanel
 from plone.registry.interfaces import IRegistry
-from ..interfaces import IShopSettings
-from ..interfaces import IShopSettingsProvider
-from .. import message_factory as _
+from bda.plone.shop.interfaces import IShopSettings
+from bda.plone.shop.interfaces import IShopSettingsProvider
+from bda.plone.shop import message_factory as _
 
 
 class ContextProxy(object):